feat: update modal component implementation
