--- src/ui/core/bottom-sheet/index.tsx
@@ -1,1 +1,0 @@
-export * from './backdrop';
